Extract action state-tracking out of caches.ts
